+++ flow/scenarios/base_scenario.py
@@ -128,9 +128,9 @@ class Scenario(Serializable):
             ])
 
         # generate starting position for vehicles in the network
-        num_vehicles = len(vehicles.get_ids())
         kwargs = initial_config.additional_params
-        positions, lanes = self.generate_starting_positions(kwargs=kwargs)
+        positions, lanes = self.generate_starting_positions(num_vehicles = vehicles.num_vehicles,
+                                                            **kwargs)
 
         # create the sumo configuration files using the generator class
         cfg_name = self.generator.generate_cfg(self.net_params,
