+++ flow/scenarios/highway/scenario.py
@@ -49,3 +49,26 @@ class HighwayScenario(Scenario):
         """See parent class."""
         edgestarts = [("highway_0", 0)]
         return edgestarts
+
+    def gen_custom_start_pos(self, initial_config, num_vehicles, **kwargs):
+        """Generate a user defined set of starting positions.
+        This method is just used for testing.
+
+        Parameters
+        ----------
+        initial_config : InitialConfig type
+            see flow/core/params.py
+        num_vehicles : int
+            number of vehicles to be placed on the network
+        kwargs : dict
+            extra components, usually defined during reset to overwrite initial
+            config parameters
+
+        Returns
+        -------
+        startpositions : list of tuple (float, float)
+            list of start positions [(edge0, pos0), (edge1, pos1), ...]
+        startlanes : list of int
+            list of start lanes
+        """
+        return kwargs["start_positions"], kwargs["start_lanes"]
